1.6

+++ src/main/java/com/leng/jadefine/controller/system/UserController.java
@@ -4,7 +4,6 @@ import com.leng.jadefine.conmmon.JsonResult;
 import com.leng.jadefine.model.Order;
 import com.leng.jadefine.model.Product;
 import com.leng.jadefine.model.User;
-import com.leng.jadefine.service.impl.AdminServiceImpl;
 import com.leng.jadefine.service.impl.OrderServiceImpl;
 import com.leng.jadefine.service.impl.UserServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,11 +14,9 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 
-import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 
 /**
  * FileName:UserController
@@ -67,12 +64,6 @@ public class UserController {
     @PostMapping({"save","update"})
     @ResponseBody
     public JsonResult from(@Valid User user){
-        User user1=userService.queryByUserName(user.getUserName());
-        System.out.println(user1);
-        if(user1!=null){
-            System.out.println("用户名重复");
-            return JsonResult.error("用户名重复");
-        }
         System.out.println(user);
         userService.saveUser(user);
         return JsonResult.success();
@@ -81,8 +72,13 @@ public class UserController {
     @GetMapping("delete")
     @ResponseBody
     public JsonResult delete(int id){
+        System.out.println(id);
         User user = userService.queryById(id);
         if(user!=null){
+            List<Order> orderList = orderService.queryByUserId(user.getId());
+            if(orderList!=null){
+                return JsonResult.error("该用户存在未完成订单，不可删除");
+            }
             userService.deleteUser(id);
             return JsonResult.success();
         }else {
